g_local.h: Fix compilation with -Wall on gcc 4.4.4 x86_64

+++ g_local.h
@@ -1,5 +1,6 @@
 // g_local.h -- local definitions for game module
 
+#include <stdint.h>
 #include "q_shared.h"
 
 // define GAME_INCLUDE so that game.h does not define the
@@ -492,10 +493,10 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define	FOFS(x) (int)&(((edict_t *)0)->x)
-#define	STOFS(x) (int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (int)&(((gclient_t *)0)->x)
+#define	FOFS(x) (uintptr_t)&(((edict_t *)0)->x)
+#define	STOFS(x) (uintptr_t)&(((spawn_temp_t *)0)->x)
+#define	LLOFS(x) (uintptr_t)&(((level_locals_t *)0)->x)
+#define	CLOFS(x) (uintptr_t)&(((gclient_t *)0)->x)
 
 #define random()	((rand () & 0x7fff) / ((float)0x7fff))
 #define crandom()	(2.0 * (random() - 0.5))
